beizuri-print-server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyusb==1.2.1